+++ ravenclaw/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-__version__ = '0.18.3'
+__version__ = '0.18.4'
 
 __all__ = ['__version__']